rename module (see review https://github.com/SwissGacho/Haushaltbuch/pull/53/files/48b7bb500f48b416d2e43b44150be5d80a4f9e61#diff-084ffdb8fd9771036f2d6699ee6fc0bc202d4774f1176577207b7b18761bddd5)
